remove Breakpoints_v2.xcbkptlist
不要なファイルを削除しました
